alters tests back to mirroring SRC file structure
